pyproject.toml
[project]
name = "unknown-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.7.0",
    "botbuilder-schema>=4.17.0",
    "chainlit>=2.5.5",
    "crawl4ai>=0.6.3",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.13",
    "google-genai>=1.19.0",
    "google-generativeai>=0.8.5",
    "huggingface-hub>=0.32.4",
    "langchain>=0.3.26",
    "matplotlib>=3.10.3",
    "mcp>=1.9.3",
    "openai>=1.90.0",
    "playwright>=1.52.0",
    "pymysql>=1.1.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "seaborn>=0.13.2",
    "semantic-kernel[google,hugging-face,mcp]>=1.28.1",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.45.1",
    "streamlit-markdown>=1.1.0",
    "streamlit-mermaid>=0.3.0",
    "transformers>=4.52.4",
    "uvicorn>=0.34.3",
]
